Move extractor functions to pipelines module
