+++ magenta/models/music_vae/data.py
@@ -24,14 +24,17 @@ import copy
 import functools
 import itertools
 
+import numpy as np
+import tensorflow as tf
+import tensorflow_datasets as tfds
+
 import magenta.music as mm
+import magenta.pipelines.drum_pipelines
+import magenta.pipelines.melody_pipelines
 from magenta.music import chords_lib
 from magenta.music import drums_encoder_decoder
 from magenta.music import sequences_lib
 from magenta.protobuf import music_pb2
-import numpy as np
-import tensorflow as tf
-import tensorflow_datasets as tfds
 
 PIANO_MIN_MIDI_PITCH = 21
 PIANO_MAX_MIDI_PITCH = 108
@@ -711,7 +714,7 @@ class OneHotMelodyConverter(LegacyEventListOneHotConverter):
       return mm.Melody(
           steps_per_bar=steps_per_bar, steps_per_quarter=steps_per_quarter)
     melody_extractor_fn = functools.partial(
-        mm.extract_melodies,
+        magenta.pipelines.melody_pipelines.extract_melodies,
         min_bars=1,
         gap_bars=gap_bars or float('inf'),
         max_steps_truncate=max_steps_truncate,
@@ -795,7 +798,7 @@ class DrumsConverter(BaseNoteSequenceConverter):
     self._roll_output = roll_output
 
     self._drums_extractor_fn = functools.partial(
-        mm.extract_drum_tracks,
+        magenta.pipelines.drum_pipelines.extract_drum_tracks,
         min_bars=1,
         gap_bars=gap_bars or float('inf'),
         max_steps_truncate=self._steps_per_bar * max_bars if max_bars else None,
