+++ magenta/models/pianoroll_rnn_nade/pianoroll_rnn_nade_pipeline.py
@@ -15,7 +15,8 @@
 """Pipeline to create Pianoroll RNN-NADE dataset."""
 
 import magenta.music as mm
-from magenta.pipelines import dag_pipeline
+from magenta.music import sequences_lib, PianorollSequence
+from magenta.pipelines import dag_pipeline, statistics
 from magenta.pipelines import event_sequence_pipeline
 from magenta.pipelines import note_sequence_pipelines
 from magenta.pipelines import pipeline
@@ -35,7 +36,7 @@ class PianorollSequenceExtractor(pipeline.Pipeline):
     self._max_steps = max_steps
 
   def transform(self, quantized_sequence):
-    pianoroll_seqs, stats = mm.extract_pianoroll_sequences(
+    pianoroll_seqs, stats = extract_pianoroll_sequences(
         quantized_sequence,
         min_steps_discard=self._min_steps,
         max_steps_truncate=self._max_steps)
@@ -86,3 +87,79 @@ def get_pipeline(config, min_steps, max_steps, eval_ratio):
     dag[dag_pipeline.DagOutput(mode + '_pianoroll_tracks')] = encoder_pipeline
 
   return dag_pipeline.DAGPipeline(dag)
+
+
+def extract_pianoroll_sequences(
+    quantized_sequence, start_step=0, min_steps_discard=None,
+    max_steps_discard=None, max_steps_truncate=None):
+  """Extracts a polyphonic track from the given quantized NoteSequence.
+
+  Currently, this extracts only one pianoroll from a given track.
+
+  Args:
+    quantized_sequence: A quantized NoteSequence.
+    start_step: Start extracting a sequence at this time step. Assumed
+        to be the beginning of a bar.
+    min_steps_discard: Minimum length of tracks in steps. Shorter tracks are
+        discarded.
+    max_steps_discard: Maximum length of tracks in steps. Longer tracks are
+        discarded. Mutually exclusive with `max_steps_truncate`.
+    max_steps_truncate: Maximum length of tracks in steps. Longer tracks are
+        truncated. Mutually exclusive with `max_steps_discard`.
+
+  Returns:
+    pianoroll_seqs: A python list of PianorollSequence instances.
+    stats: A dictionary mapping string names to `statistics.Statistic` objects.
+
+  Raises:
+    ValueError: If both `max_steps_discard` and `max_steps_truncate` are
+        specified.
+  """
+
+  if (max_steps_discard, max_steps_truncate).count(None) == 0:
+    raise ValueError(
+        'Only one of `max_steps_discard` and `max_steps_truncate` can be '
+        'specified.')
+  sequences_lib.assert_is_relative_quantized_sequence(quantized_sequence)
+
+  stats = dict((stat_name, statistics.Counter(stat_name)) for stat_name in
+               ['pianoroll_tracks_truncated_too_long',
+                'pianoroll_tracks_discarded_too_short',
+                'pianoroll_tracks_discarded_too_long',
+                'pianoroll_tracks_discarded_more_than_1_program'])
+
+  steps_per_bar = sequences_lib.steps_per_bar_in_quantized_sequence(
+      quantized_sequence)
+
+  # Create a histogram measuring lengths (in bars not steps).
+  stats['pianoroll_track_lengths_in_bars'] = statistics.Histogram(
+      'pianoroll_track_lengths_in_bars',
+      [0, 1, 10, 20, 30, 40, 50, 100, 200, 500, 1000])
+
+  # Allow only 1 program.
+  programs = set()
+  for note in quantized_sequence.notes:
+    programs.add(note.program)
+  if len(programs) > 1:
+    stats['pianoroll_tracks_discarded_more_than_1_program'].increment()
+    return [], stats.values()
+
+  # Translate the quantized sequence into a PianorollSequence.
+  pianoroll_seq = PianorollSequence(quantized_sequence=quantized_sequence,
+                                    start_step=start_step)
+
+  pianoroll_seqs = []
+  num_steps = pianoroll_seq.num_steps
+
+  if min_steps_discard is not None and num_steps < min_steps_discard:
+    stats['pianoroll_tracks_discarded_too_short'].increment()
+  elif max_steps_discard is not None and num_steps > max_steps_discard:
+    stats['pianoroll_tracks_discarded_too_long'].increment()
+  else:
+    if max_steps_truncate is not None and num_steps > max_steps_truncate:
+      stats['pianoroll_tracks_truncated_too_long'].increment()
+      pianoroll_seq.set_length(max_steps_truncate)
+    pianoroll_seqs.append(pianoroll_seq)
+    stats['pianoroll_track_lengths_in_bars'].increment(
+        num_steps // steps_per_bar)
+  return pianoroll_seqs, stats.values()
