+++ magenta/models/pianoroll_rnn_nade/pianoroll_rnn_nade_sequence_generator.py
@@ -17,6 +17,7 @@
 import functools
 
 from magenta.models.pianoroll_rnn_nade import pianoroll_rnn_nade_model
+from magenta.models.pianoroll_rnn_nade import pianoroll_rnn_nade_pipeline
 import magenta.music as mm
 
 
@@ -87,7 +88,7 @@ class PianorollRnnNadeSequenceGenerator(mm.BaseSequenceGenerator):
     quantized_primer_sequence = mm.quantize_note_sequence(
         primer_sequence, self.steps_per_quarter)
 
-    extracted_seqs, _ = mm.extract_pianoroll_sequences(
+    extracted_seqs, _ = pianoroll_rnn_nade_pipeline.extract_pianoroll_sequences(
         quantized_primer_sequence, start_step=input_start_step)
     assert len(extracted_seqs) <= 1
