+++ magenta/models/rl_tuner/note_rnn_loader.py
@@ -34,10 +34,10 @@ These functions are necessary for use with the RL Tuner class.
 import os
 
 import magenta
+import magenta.pipelines.melody_pipelines
 from magenta.common import sequence_example_lib
 from magenta.models.rl_tuner import rl_tuner_ops
 from magenta.models.shared import events_rnn_graph
-from magenta.music import melodies_lib
 from magenta.music import midi_io
 from magenta.music import sequences_lib
 import numpy as np
@@ -294,9 +294,8 @@ class NoteRNNLoader(object):
     self.primer_sequence = midi_io.midi_file_to_sequence_proto(self.midi_primer)
     quantized_seq = sequences_lib.quantize_note_sequence(
         self.primer_sequence, steps_per_quarter=4)
-    extracted_melodies, _ = melodies_lib.extract_melodies(quantized_seq,
-                                                          min_bars=0,
-                                                          min_unique_pitches=1)
+    extracted_melodies, _ = magenta.pipelines.melody_pipelines.extract_melodies(
+        quantized_seq, min_bars=0, min_unique_pitches=1)
     self.primer = extracted_melodies[0]
     self.steps_per_bar = self.primer.steps_per_bar
