+++ magenta/music/__init__.py
@@ -35,15 +35,12 @@ from magenta.music.chords_encoder_decoder import TriadChordOneHotEncoding
 
 from magenta.music.chords_lib import BasicChordRenderer
 from magenta.music.chords_lib import ChordProgression
-from magenta.music.chords_lib import extract_chords
-from magenta.music.chords_lib import extract_chords_for_melodies
 
 from magenta.music.constants import *  # pylint: disable=wildcard-import
 
 from magenta.music.drums_encoder_decoder import MultiDrumOneHotEncoding
 
 from magenta.music.drums_lib import DrumTrack
-from magenta.music.drums_lib import extract_drum_tracks
 from magenta.music.drums_lib import midi_file_to_drum_track
 
 from magenta.music.encoder_decoder import ConditionalEventSequenceEncoderDecoder
@@ -57,11 +54,9 @@ from magenta.music.encoder_decoder import OptionalEventSequenceEncoder
 
 from magenta.music.events_lib import NonIntegerStepsPerBarError
 
-from magenta.music.lead_sheets_lib import extract_lead_sheet_fragments
 from magenta.music.lead_sheets_lib import LeadSheet
 
 from magenta.music.melodies_lib import BadNoteError
-from magenta.music.melodies_lib import extract_melodies
 from magenta.music.melodies_lib import Melody
 from magenta.music.melodies_lib import midi_file_to_melody
 from magenta.music.melodies_lib import PolyphonicMelodyError
@@ -101,13 +96,11 @@ from magenta.music.performance_encoder_decoder import NotePerformanceEventSequen
 from magenta.music.performance_encoder_decoder import PerformanceModuloEncoding
 from magenta.music.performance_encoder_decoder import PerformanceOneHotEncoding
 
-from magenta.music.performance_lib import extract_performances
 from magenta.music.performance_lib import MetricPerformance
 from magenta.music.performance_lib import Performance
 
 from magenta.music.pianoroll_encoder_decoder import PianorollEncoderDecoder
 
-from magenta.music.pianoroll_lib import extract_pianoroll_sequences
 from magenta.music.pianoroll_lib import PianorollSequence
