+++ magenta/pipelines/chord_pipelines_test.py
@@ -13,12 +13,12 @@
 # limitations under the License.
 
 """Tests for chord_pipelines."""
-
+import magenta.pipelines.melody_pipelines
 from magenta.common import testing_lib as common_testing_lib
 from magenta.music import chords_lib
 from magenta.music import constants
 from magenta.music import sequences_lib
-from magenta.music import testing_lib
+from magenta.music import testing_lib as music_testing_lib
 from magenta.pipelines import chord_pipelines
 from magenta.protobuf import music_pb2
 import tensorflow as tf
@@ -27,6 +27,20 @@ NO_CHORD = constants.NO_CHORD
 
 
 class ChordPipelinesTest(tf.test.TestCase):
+
+  def setUp(self):
+    self.steps_per_quarter = 1
+    self.note_sequence = music_testing_lib.parse_test_proto(
+        music_pb2.NoteSequence,
+        """
+        time_signatures: {
+          numerator: 4
+          denominator: 4
+        }
+        tempos: {
+          qpm: 60
+        }
+        """)
 
   def _unit_transform_test(self, unit, input_instance,
                            expected_outputs):
@@ -46,7 +60,7 @@ class ChordPipelinesTest(tf.test.TestCase):
           denominator: 4}
         tempos: {
           qpm: 60}""")
-    testing_lib.add_chords_to_sequence(
+    music_testing_lib.add_chords_to_sequence(
         note_sequence, [('C', 2), ('Am', 4), ('F', 5)])
     quantized_sequence = sequences_lib.quantize_note_sequence(
         note_sequence, steps_per_quarter=1)
@@ -60,6 +74,80 @@ class ChordPipelinesTest(tf.test.TestCase):
     unit = chord_pipelines.ChordsExtractor(all_transpositions=False)
     self._unit_transform_test(unit, quantized_sequence,
                               expected_chord_progressions)
+
+  def testExtractChords(self):
+    music_testing_lib.add_chords_to_sequence(
+        self.note_sequence, [('C', 2), ('G7', 6), ('F', 8)])
+    quantized_sequence = sequences_lib.quantize_note_sequence(
+        self.note_sequence, self.steps_per_quarter)
+    quantized_sequence.total_quantized_steps = 10
+    chord_progressions, _ = chord_pipelines.extract_chords(quantized_sequence)
+    expected = [[NO_CHORD, NO_CHORD, 'C', 'C', 'C', 'C', 'G7', 'G7', 'F', 'F']]
+    self.assertEqual(expected, [list(chords) for chords in chord_progressions])
+
+  def testExtractChordsAllTranspositions(self):
+    music_testing_lib.add_chords_to_sequence(
+        self.note_sequence, [('C', 1)])
+    quantized_sequence = sequences_lib.quantize_note_sequence(
+        self.note_sequence, self.steps_per_quarter)
+    quantized_sequence.total_quantized_steps = 2
+    chord_progressions, _ = chord_pipelines.extract_chords(
+        quantized_sequence, all_transpositions=True)
+    expected = list(zip([NO_CHORD] * 12, ['Gb', 'G', 'Ab', 'A', 'Bb', 'B',
+                                          'C', 'Db', 'D', 'Eb', 'E', 'F']))
+    self.assertEqual(expected, [tuple(chords) for chords in chord_progressions])
+
+  def testExtractChordsForMelodies(self):
+    music_testing_lib.add_track_to_sequence(
+        self.note_sequence, 0,
+        [(12, 100, 2, 4), (11, 1, 6, 11)])
+    music_testing_lib.add_track_to_sequence(
+        self.note_sequence, 1,
+        [(12, 127, 2, 4), (14, 50, 6, 8),
+         (50, 100, 33, 37), (52, 100, 34, 37)])
+    music_testing_lib.add_chords_to_sequence(
+        self.note_sequence,
+        [('C', 2), ('G7', 6), ('Cmaj7', 33)])
+    quantized_sequence = sequences_lib.quantize_note_sequence(
+        self.note_sequence, self.steps_per_quarter)
+
+    melodies, _ = magenta.pipelines.melody_pipelines.extract_melodies(
+        quantized_sequence, min_bars=1, gap_bars=2, min_unique_pitches=2,
+        ignore_polyphonic_notes=True)
+    chord_progressions, _ = chord_pipelines.extract_chords_for_melodies(
+        quantized_sequence, melodies)
+    expected = [[NO_CHORD, NO_CHORD, 'C', 'C', 'C', 'C',
+                 'G7', 'G7', 'G7', 'G7', 'G7'],
+                [NO_CHORD, NO_CHORD, 'C', 'C', 'C', 'C', 'G7', 'G7'],
+                ['G7', 'Cmaj7', 'Cmaj7', 'Cmaj7', 'Cmaj7']]
+    self.assertEqual(expected, [list(chords) for chords in chord_progressions])
+
+  def testExtractChordsForMelodiesCoincidentChords(self):
+    music_testing_lib.add_track_to_sequence(
+        self.note_sequence, 0,
+        [(12, 100, 2, 4), (11, 1, 6, 11)])
+    music_testing_lib.add_track_to_sequence(
+        self.note_sequence, 1,
+        [(12, 127, 2, 4), (14, 50, 6, 8),
+         (50, 100, 33, 37), (52, 100, 34, 37)])
+    music_testing_lib.add_chords_to_sequence(
+        self.note_sequence,
+        [('C', 2), ('G7', 6), ('E13', 8), ('Cmaj7', 8)])
+    quantized_sequence = sequences_lib.quantize_note_sequence(
+        self.note_sequence, self.steps_per_quarter)
+
+    melodies, _ = magenta.pipelines.melody_pipelines.extract_melodies(
+        quantized_sequence, min_bars=1, gap_bars=2, min_unique_pitches=2,
+        ignore_polyphonic_notes=True)
+    chord_progressions, stats = chord_pipelines.extract_chords_for_melodies(
+        quantized_sequence, melodies)
+    expected = [[NO_CHORD, NO_CHORD, 'C', 'C', 'C', 'C', 'G7', 'G7'],
+                ['Cmaj7', 'Cmaj7', 'Cmaj7', 'Cmaj7', 'Cmaj7']]
+    stats_dict = dict((stat.name, stat) for stat in stats)
+    self.assertIsNone(chord_progressions[0])
+    self.assertEqual(expected,
+                     [list(chords) for chords in chord_progressions[1:]])
+    self.assertEqual(stats_dict['coincident_chords'].count, 1)
 
 
 if __name__ == '__main__':
